spelling and move bodies
